move custom components to tina folder
