+++ tina/config.tsx
@@ -3,7 +3,7 @@ import { contentBlockSchema } from "../components/blocks/content";
 import { featureBlockSchema } from "../components/blocks/features";
 import { heroBlockSchema } from "../components/blocks/hero";
 import { testimonialBlockSchema } from "../components/blocks/testimonial";
-import { ColorPickerInput } from "../components/fields/color";
+import { ColorPickerInput } from "./fields/color";
 import { iconSchema } from "../components/util/icon";
 
 const config = defineConfig({
